A saved world loads on Postgres 14+, where EXTRACT(EPOCH) is NUMERIC
Since PostgreSQL 14, EXTRACT returns NUMERIC rather than double precision,
so `load_world` and `list_worlds` panicked decoding `updated_at_ms` as f64
("mismatched types; Rust type f64 is not compatible with SQL type NUMERIC").
The Router calls list_worlds at boot, so on any modern Postgres with a saved
world no dive could be created. Cast to float8 in both queries.

+++ server/crates/meld-db/src/lib.rs
@@ -5732,7 +5732,7 @@ impl Db {
             Backend::Pg(pool) => {
                 let row = sqlx::query(
                     "SELECT seed, tick_count, shift_generation, sections, delta,
-                            EXTRACT(EPOCH FROM updated_at) * 1000 AS updated_at_ms
+                            (EXTRACT(EPOCH FROM updated_at) * 1000)::float8 AS updated_at_ms
                      FROM worlds WHERE world_key = $1",
                 )
                 .bind(world_key)
@@ -5763,7 +5763,7 @@ impl Db {
             Backend::Pg(pool) => {
                 let rows = sqlx::query(
                     "SELECT world_key, seed, tick_count, shift_generation, sections, delta,
-                            EXTRACT(EPOCH FROM updated_at) * 1000 AS updated_at_ms
+                            (EXTRACT(EPOCH FROM updated_at) * 1000)::float8 AS updated_at_ms
                      FROM worlds ORDER BY world_key",
                 )
                 .fetch_all(pool)
